chore: transform workers to active job and cleanup sidekiq
