+++ app/jobs/cleanup/base_job.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Cleanup
+  class BaseJob < ::ApplicationJob
+    queue_as :cleanup
+    sidekiq_options retry: false
+  end
+end
