+++ app/jobs/loaders/base_job.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Loaders
+  class BaseJob < ::ApplicationJob
+    queue_as :loaders
+    sidekiq_options retry: false
+  end
+end
