+++ app/jobs/loaders/model_job.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'rsi/models_loader'
+
+module Loaders
+  class ModelJob < ::Loaders::BaseJob
+    def perform(rsi_id)
+      ::Rsi::ModelsLoader
+        .new(vat_percent: Rails.application.secrets[:rsi_vat_percent])
+        .one(rsi_id)
+    end
+  end
+end
