+++ app/jobs/notifications/base_job.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Notifications
+  class BaseJob < ::ApplicationJob
+    queue_as :notifications
+    sidekiq_options retry: true
+  end
+end
