--- app/workers/roadmap_worker.rb
@@ -1,27 +1,0 @@
-# frozen_string_literal: true
-
-require 'rsi/roadmap_loader'
-require 'discord/roadmap_update'
-
-class RoadmapWorker
-  include Sidekiq::Worker
-  sidekiq_options retry: false, queue: (ENV['ROADMAP_LOADER_QUEUE'] || 'fleetyards_roadmap_loader').to_sym
-
-  def perform
-    count_before = PaperTrail::Version.where(item_type: 'RoadmapItem').count
-
-    ::Rsi::RoadmapLoader.new.fetch
-
-    changes = PaperTrail::Version.where(item_type: 'RoadmapItem').count - count_before
-
-    return if changes.zero?
-
-    RoadmapMailer.notify_admin(changes).deliver_later
-
-    Discord::RoadmapUpdate.new(changes: changes).run
-
-    ActionCable.server.broadcast('roadmap', {
-      changes: changes
-    }.to_json)
-  end
-end
